Add query escaping for string values
This commit does a few things:

- Fixes all unit tests via:
  - Setting named scope conditionally
  - Only adds dotescaped values if value is not nil
  - Only includes parent opts if using dotnumbered option

Makes unit tests easier to debug by adding a `name` variable to each
test-case, so a developer doesn't have to count the index of a failing
test case.

Finally, patch adds the `escaped` option to allow a user to get URL
Query escaped values for string parameters if the `escaped` option is
supplied, and `value.Kind()` is `string`.